refactor(vibe): reorganise components
