+++ ui/apps/moderation/src/components/dashboard/tabs/general/max-applicants.tsx
@@ -5,7 +5,7 @@ import Text from '@akashaorg/design-system-core/lib/components/Text';
 
 import TextField from '@akashaorg/design-system-core/lib/components/TextField';
 
-import { PageHeader, IPageHeaderProps } from '../common';
+import { IPageHeaderProps, PageHeader } from '../../../common';
 
 export interface IEditMaxApplicantsProps extends IPageHeaderProps {
   introLabel: string;
