+++ ui/apps/moderation/src/components/dashboard/tabs/general/resign-confirmation.tsx
