+++ ui/apps/moderation/src/components/dashboard/tabs/general/resign-role.tsx
@@ -4,7 +4,7 @@ import Box from '@akashaorg/design-system-core/lib/components/Box';
 import Text from '@akashaorg/design-system-core/lib/components/Text';
 import TextField from '@akashaorg/design-system-core/lib/components/TextField';
 
-import { PageHeader, IPageHeaderProps } from '../common';
+import { IPageHeaderProps, PageHeader } from '../../../common';
 
 export interface IResignRoleProps extends IPageHeaderProps {
   textLine1Label: string;
